Projservice/feature/add_remove_product_from_cart

+++ projservice/storegoservice/src/main/java/com/storego/storegoservice/services/StoreServices.java
@@ -3,6 +3,7 @@ import com.storego.storegoservice.model.*;
 import com.storego.storegoservice.repository.CartProductRepository;
 import com.storego.storegoservice.repository.CartRepository;
 import com.storego.storegoservice.repository.ProductRepository;
+import org.springframework.expression.ExpressionException;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Service;
 
@@ -12,10 +13,9 @@ import java.util.HashSet;
 
 // Connection to DB
 import org.springframework.beans.factory.annotation.Autowired;
-import com.storego.storegoservice.exception.ResourceNotFoundException;
+
 import com.storego.storegoservice.repository.PersonRepository;
 
-import javax.persistence.EntityNotFoundException;
 
 @Service
 public class StoreServices {
@@ -29,56 +29,73 @@ public class StoreServices {
     @Autowired
     private CartProductRepository cartProductRepository;
 
-    private Set<Person> clientsInStore;
+    @Autowired
+    private CartRepository cartRepository;
+
 
     public StoreServices() {
-        this.clientsInStore = new HashSet<>();
-    }
 
-    public Set<Person> getClientsInStore() {
-        return clientsInStore;
     }
 
     public void enterStore(Long nif){
         Person p = personRepository.findByNif(nif);
+        Cart c = new Cart(p);
+        cartRepository.save(c);
         p.setLast_visit(new Date());
         personRepository.save(p);
-        clientsInStore.add(p);
     }
 
     public void leaveStore(Long nif){
-        Person p = personRepository.findByNif(nif);
-        System.out.println(p.getName());
-        clientsInStore.remove(p);
+        Set<CartProduct> products = cartProductRepository.findByCartPersonNif(nif);
+        for(CartProduct p: products){
+            cartProductRepository.delete(p);
+        }
+        Cart c = cartRepository.findByPersonNif(nif);
+        cartRepository.delete(c);
     }
 
-    public void removeProduct(Long nif, Long prod_id, Integer quantity) {
-        CartProduct cp = cartProductRepository.findByCartIdAndProductId(nif, prod_id);
-        Integer units = cp.getUnits();
-        if (units > quantity){
-            cp.setUnits(units-quantity);
-            cartProductRepository.save(cp);
+    public void removeProductFromCart(Long nif, Long prod_id, Integer quantity) throws Exception{
+        CartProduct cp = cartProductRepository.findByCartPersonNifAndProductId(nif, prod_id);
+
+        Product product = productRepository.findById(prod_id).orElseThrow(() -> new Exception("Product not found!"));
+
+        if (cp != null) {
+            Integer units = cp.getUnits();
+            if (units > quantity) {
+                cp.setUnits(units - quantity);
+                cartProductRepository.save(cp);
+            } else {
+                cartProductRepository.delete(cp);
+            }
+            Integer stock = product.getStock_current();
+            product.setStock_current(stock - quantity);
         } else {
-            cartProductRepository.delete(cp);
+            throw(new Exception("User hasn't got that product!"));
         }
-        Product p = productRepository.findById(prod_id).orElseThrow(() -> new EntityNotFoundException("Product not found!"));
-        Integer stock = p.getStock_current();
-        p.setStock_current(stock - quantity);
     }
 
-    //Implement later
-    public void addProduct(Long nif, Long prod_id, Integer quantity) {
-        CartProduct cp = cartProductRepository.findByCartIdAndProductId(nif, prod_id);
-        Integer units = cp.getUnits();
-        if (units > quantity){
-            cp.setUnits(units-quantity);
+    public void addProductToCart(Long nif, Long prod_id, Integer quantity) throws Exception{
+        CartProduct cp = cartProductRepository.findByCartPersonNifAndProductId(nif, prod_id);
+        Product product = productRepository.findById(prod_id).orElseThrow(() -> new Exception("Product not found!"));
+
+        if (cp == null) {
+            Cart cart = cartRepository.findByPersonNif(nif);
+            if (cart == null) throw(new Exception("Cart of client not found!"));
+            cp = new CartProduct(cart, product, quantity);
             cartProductRepository.save(cp);
+
         } else {
-            cartProductRepository.delete(cp);
+            Integer cp_units = cp.getUnits();
+            Integer stock = product.getStock_current();
+            if (stock >= cp_units + quantity) {
+                cp.setUnits(cp_units + quantity);
+                cartProductRepository.save(cp);
+                product.setStock_current(stock - quantity);
+                productRepository.save(product);
+            } else {
+                throw(new Exception("Stock is not enough!"));
+            }
         }
-        Product p = productRepository.findById(prod_id).orElseThrow(() -> new EntityNotFoundException("Product not found!"));
-        Integer stock = p.getStock_current();
-        p.setStock_current(stock - quantity);
     }
 
 }
